requirements.txt
flask~=3.0.0
python-dotenv~=1.0.0
storage3~=0.1.0
httpx>=0.13,<0.26
postgrest-py>=0.2.0,<0.11
scikit-learn~=1.3.2
pymongo~=4.5.0
pandas~=2.1.2
supabase~=2.0.1